requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0
django-annoying==0.10.4
django-dbbackup==3.2.0
idna==2.6
Pillow==5.0.0
psycopg2==2.7.3.2
python-resize-image==1.1.11
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
